+++ client/src/App.js
@@ -1,0 +1,5 @@
+function App() {
+  return <div>Hello world</div>;
+}
+
+export default App;
