+++ client/src/redux/reducer.js
@@ -1,0 +1,12 @@
+const initState = {
+  darkMode: false,
+  connect: {},
+  notification: {},
+  game: {},
+};
+
+const reducer = (state = initState, action) => {
+  return state;
+};
+
+export default reducer;
